enemymove script
