--- Robachinos/Assets/Scripts/Enemy/enemywaypoints.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6d147de6a60912d4498ba318cc13321a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
